base-item: Don't read guint64 properties as gsize
It's probably more portable to be pedantic about such things, and even
if it isn't, avoiding hairy assumptions is always good.

+++ src/photos-base-item.c
@@ -858,7 +858,7 @@ photos_base_item_guess_save_sizes_from_buffer (GeglBuffer *buffer,
   GeglNode *buffer_source;
   GeglNode *graph;
   GeglNode *guess_sizes;
-  gsize sizes[2];
+  guint64 sizes[2];
 
   graph = gegl_node_new ();
   buffer_source = gegl_node_new_child (graph, "operation", "gegl:buffer-source", "buffer", buffer, NULL);
@@ -883,9 +883,9 @@ photos_base_item_guess_save_sizes_from_buffer (GeglBuffer *buffer,
 
   gegl_node_get (guess_sizes, "size", &sizes[0], "size-1", &sizes[1], NULL);
   if (out_full_size != NULL)
-    *out_full_size = sizes[0];
+    *out_full_size = (gsize) sizes[0];
   if (out_reduced_size != NULL)
-    *out_reduced_size = sizes[1];
+    *out_reduced_size = (gsize) sizes[1];
 
   g_object_unref (graph);
 }
